requirements.txt
opencv-python==4.11.0.86
PyQt5==5.15.9
ultralytics==8.3.93
pyserial==3.5
pillow==10.4.0